+++ src/js/core/accordion.js
@@ -1,5 +1,5 @@
 import { Class, Toggable } from '../mixin/index';
-import { $, getIndex, toJQuery, Transition } from '../util/index';
+import { $, getIndex, toJQuery } from '../util/index';
 
 export default function (UIkit) {
 
@@ -87,7 +87,7 @@ export default function (UIkit) {
 
                     el = $(el);
 
-                    var content = el.find(this.content), isItem = el.is(item), state = isItem && !el.hasClass(this.clsOpen);
+                    var isItem = el.is(item), state = isItem && !el.hasClass(this.clsOpen);
 
                     if (!state && isItem && !this.collapsible && active.length < 2) {
                         return;
@@ -95,18 +95,21 @@ export default function (UIkit) {
 
                     el.toggleClass(this.clsOpen, state);
 
-                    if (!Transition.inProgress(content.parent())) {
-                        content.wrap('<div>').parent().attr('hidden', state);
+                    var content = el[0]._wrapper ? el[0]._wrapper.children().first() : el.find(this.content);
+
+                    if (!el[0]._wrapper) {
+                        el[0]._wrapper = content.wrap('<div>').parent().attr('hidden', state);
                     }
 
                     this._toggleImmediate(content, true);
-                    this.toggleElement(content.parent(), state, animate).then(() => {
+                    this.toggleElement(el[0]._wrapper, state, animate).then(() => {
                         if (el.hasClass(this.clsOpen) === state) {
 
                             if (!state) {
                                 this._toggleImmediate(content, false);
                             }
 
+                            el[0]._wrapper = null;
                             content.unwrap();
                         }
                     });